requirements.txt
torch
torchvision
monai
nibabel
numpy
pillow
imageio
tqdm
transformers
timm
matplotlib
